requirements.txt
ics==0.7.2
python-dotenv==0.21.0
python_dateutil==2.8.2
selenium==4.14.0
webdriver_manager==4.0.1
packaging==21.3
Jinja2==3.1.2
python-slugify==7.0.0
gitpython==3.1.40
